delete picture with object if not std picture

+++ branches/Inmaa_Admin/src/com/inmaa/admin/tools/Utils.java
@@ -68,17 +68,20 @@ public class Utils {
 	}
 
 	public static void deletePicture(String pictureName) {
-		File file = new File(ConfigBean.getImgFilePath() +"/"+ pictureName);
-		Path path = file.toPath();
-		try {
-			Files.delete(path);
-		} catch (NoSuchFileException x) {
-			System.err.format("%s: no such" + " file or directory%n", path);
-		} catch (DirectoryNotEmptyException x) {
-			System.err.format("%s not empty%n", path);
-		} catch (IOException x) {
-			// File permission problems are caught here.
-			System.err.println(x);
+		if(pictureName.startsWith("img"))
+		{
+			File file = new File(ConfigBean.getImgFilePath() +"/"+ pictureName);
+			Path path = file.toPath();
+			try {
+				Files.delete(path);
+			} catch (NoSuchFileException x) {
+				System.err.format("%s: no such" + " file or directory%n", path);
+			} catch (DirectoryNotEmptyException x) {
+				System.err.format("%s not empty%n", path);
+			} catch (IOException x) {
+				// File permission problems are caught here.
+				System.err.println(x);
+			}			
 		}
 	}
